Completely wipe devices by default
We don't want users to write an image to the disk and give it to someone, only
for that person to dump the disk image and find the personal data that might be
stored on the 'unused' space after the ISO data.

Make this configurable for people who know the drives are indeed blank.