Queue the device writes according to what root hub they are connected
Writing to 10 devices on root hub 1 and another 10 on root hub 2 is much faster
than writing 20 devices on each hub in turn. This is due to the much larger
throughput of the PCIe bus compared with a single USB 2.0 bus.

This also means the max-threads settings value is per-root-hub, so that if you
have 10 threads configured and 3 root hubs you would be writing to 30 devices
at once.